separate imageCaption from blankstate
